+++ CHANGELOG.md
@@ -16,6 +16,11 @@ This file documents all changes to Argus-frontend. This file is primarily meant 
 - Rename the label for the severity level selector from "Max level" to "Max severity level"
 
 
+## fix.bulk-dialogue-fix
+
+### Fixed
+- Fix help text refering to a singular incident when ACK-ing or closing multiple incidents with the bulk operation feature.
+
 
 
 ## [Unreleased]